refactor: remove CenterCardEarlyAccess component and integrate its functionality into CenterCard
- Deleted the CenterCardEarlyAccess component to streamline the codebase.
- Integrated early access features directly into the CenterCard component, enhancing maintainability.
- Updated CSS for seamless animation transitions and improved visual engagement during loading phases.
