+++ apps/www/src/components/landing/center-card.tsx
@@ -1,52 +1,16 @@
-"use client";
-
-import { useEffect, useState } from "react";
-
 import { Icons } from "@repo/ui/components/icons";
 import { cn } from "@repo/ui/lib/utils";
 
-import { CenterCardEarlyAccess } from "./center-card-early-access";
+import { CenterCardEarlyAccessForm } from "./center-card-early-access-form";
+
+const INTRO_TEXTS = [
+  "Simplifying the way you interact with applications like Blender, Unity, Fusion360 and more.",
+  "Lightfast gives your ideas room to grow... to branch, remix and become what they're meant to be.",
+  "We integrate with your tools to make your workflow more efficient.",
+  "Ready to get early access?",
+];
 
 export const CenterCard = () => {
-  const [earlyAccessPhase, setEarlyAccessPhase] = useState(0);
-  const [loadingComplete, setLoadingComplete] = useState(false);
-
-  // Monitor CSS variables for phase changes
-  useEffect(() => {
-    const updatePhases = () => {
-      const root = document.documentElement;
-      const earlyAccess = parseFloat(
-        getComputedStyle(root).getPropertyValue("--early-access-text-phase") ||
-          "0",
-      );
-      setEarlyAccessPhase(earlyAccess);
-    };
-
-    // Set up a requestAnimationFrame loop to monitor CSS variables
-    let rafId: number;
-    const monitor = () => {
-      updatePhases();
-      rafId = requestAnimationFrame(monitor);
-    };
-
-    monitor();
-
-    return () => {
-      if (rafId) {
-        cancelAnimationFrame(rafId);
-      }
-    };
-  }, []);
-
-  // Mark loading as complete after initial animations
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setLoadingComplete(true);
-    }, 3000); // After grid lines and initial animations
-
-    return () => clearTimeout(timer);
-  }, []);
-
   return (
     <div
       className={cn(
@@ -55,12 +19,7 @@ export const CenterCard = () => {
       )}
     >
       {/* Text content (top-left, fades out during text phase) */}
-      <div
-        className={cn(
-          "center-card-text absolute top-8 right-8 left-8",
-          loadingComplete && "loading-complete",
-        )}
-      >
+      <div className="center-card-text absolute top-8 right-8 left-8">
         <p className="max-w-[400px] text-2xl font-bold">
           The intelligent creative copilot that simplifies the way you interact
           with applications like Blender, Unity, Fusion360 and more.
@@ -68,21 +27,36 @@ export const CenterCard = () => {
       </div>
 
       {/* Logo (transforms from bottom-left to center during logo phase) */}
-      <div
-        className={cn(
-          "center-card-logo absolute flex items-center justify-center",
-          loadingComplete && "loading-complete",
-        )}
-      >
+      <div className="center-card-logo absolute flex items-center justify-center">
         <Icons.logoShort className="text-primary h-12 w-12" />
       </div>
 
       {/* Early Access Chat (appears during early access phase) */}
-      <div
-        className="center-card-early-access-container absolute inset-0"
-        data-visible={earlyAccessPhase > 0.1 ? "true" : "false"}
-      >
-        <CenterCardEarlyAccess isVisible={earlyAccessPhase > 0.1} />
+      <div className="center-card-early-access-container absolute inset-0">
+        <div className="center-card-early-access absolute top-8 right-8 left-8">
+          {/* Animated intro messages - CSS driven */}
+          <div
+            className="mb-6 max-w-md space-y-4"
+            aria-live="polite"
+            aria-atomic="false"
+          >
+            {INTRO_TEXTS.map((text, index) => (
+              <div
+                key={`intro-${index}`}
+                className="early-access-text text-sm"
+                data-index={index}
+                role="status"
+              >
+                {text}
+              </div>
+            ))}
+          </div>
+
+          {/* Form area - client component */}
+          <div className="max-w-sm">
+            <CenterCardEarlyAccessForm />
+          </div>
+        </div>
       </div>
     </div>
   );
